feat(uta-client): lift trading types to uta-protocol + build SDK skeleton
Step 6 staging. Two things land together because they're co-dependent:

1) Trading types moved into `@traderalice/uta-protocol`. Files copied
   via `cp` (no byte changes) so months of refinement on BrokerError,
   IBroker, Operation, GitCommit, etc. survive intact:

     packages/uta-protocol/src/types/
       broker.ts          # ex src/domain/trading/brokers/types.ts
       git.ts             # ex src/domain/trading/git/types.ts
       contract-ext.ts    # ex src/domain/trading/contract-ext.ts
       manager.ts         # NEW — UTASummary, AggregatedEquity,
                          #       ContractSearchResult (lifted from
                          #       uta-manager.ts)

   Original locations become thin `export *` shims pointing at
   uta-protocol, so the 28 in-repo importers keep working through
   relative paths. Once domain/trading physically moves to UTA the
   shims can be dropped — internal imports already point at the right
   bytes.

   uta-protocol now depends on @traderalice/ibkr + decimal.js for the
   Contract/Decimal references inside the lifted types. The barrel
   side-effect-imports `contract-ext.ts` so the `Contract.aliceId`
   declaration merge is registered transitively.

2) UTAClient + SDK skeleton — Alice's HTTP-side adapter:

     packages/uta-protocol/src/client/UTAClient.ts
       createUTAClient({ baseUrl }) → { get, post, put, delete, request }
       UTAHttpError for typed downstream handling
       15s default timeout

     src/services/uta-client/
       UTAManagerSDK.ts    # mirrors UTAManager — async listUTAs /
                           # resolve / get / size / reconnectUTA /
                           # removeUTA / getAggregatedEquity /
                           # searchContracts
       UTAAccountSDK.ts    # mirrors UnifiedTradingAccount — async
                           # getAccount / getPositions / getOrders /
                           # getQuote / getMarketClock / log / show /
                           # status / push / reject
       index.ts            # barrel

   Setup hooks (setSnapshotHooks, setFxService, initUTA, closeAll, …)
   become no-ops on the SDK since UTA owns those concerns.

   Methods that need UTA routes which don't exist yet — commit, sync,
   simulatePriceChange, getContractDetails, exportGitState — throw
   `NotImplementedInSDK` with the route to add. Tracked for follow-up.

   The SDK is built but NOT yet wired into main.ts. That wiring +
   await cascade through consumers + physical domain/trading move is
   the next chunk.

Verified: tsc --noEmit clean, pnpm test 94 files / 1738 tests pass,
preset-catalog rename detected as 100% (no byte drift in shared code).
