+++ packages/uta-protocol/package.json
@@ -24,6 +24,8 @@
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
+    "@traderalice/ibkr": "workspace:*",
+    "decimal.js": "^10.6.0",
     "zod": "^4.3.6"
   },
   "devDependencies": {
