+++ packages/uta-protocol/src/types/index.ts
@@ -11,3 +11,11 @@
  */
 
 export * from './errors.js'
+export * from './broker.js'
+export * from './git.js'
+export * from './manager.js'
+// contract-ext extends the IBKR Contract type via declaration merge —
+// importing the package once per process suffices to register the
+// `aliceId?` field. Re-export as side-effect so consumers don't need a
+// separate import for the augmentation.
+import './contract-ext.js'
